terminal: address compile issues after merge, some review comments, move things to constants so widget can be loaded
